pyproject.toml
[project]
name = "c2sagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.4.26",
    "click==8.2.1",
    "colorama==0.4.6",
    "distro==1.9.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.115.12",
    "fastmcp==2.5.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "jiter==0.10.0",
    "markdown-it-py==3.0.0",
    "mcp==1.9.1",
    "mdurl==0.1.2",
    "openai==1.82.0",
    "openapi-pydantic==0.5.1",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pygments==2.19.1",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "rich==14.0.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sse-starlette==2.3.5",
    "starlette==0.46.2",
    "tqdm==4.67.1",
    "typer==0.16.0",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "uvicorn==0.34.2",
    "websockets==15.0.1",
    "sqlalchemy",
    "psycopg2",
    "tinydb>=4.8.2",
    "jose>=1.0.0",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.4",
    "aiohttp>=3.12.13",
    "asyncclick>=8.1.8",
    "a2a-sdk>=0.2.10",
    "jinja2>=3.1.6",
]
